Cleaning up node binding
